Moved `Controller` to abstract_ui Package
